IOT/requirements.txt
fastapi
uvicorn
pydantic
pandas
numpy
joblib
scikit-learn
imbalanced-learn
pyserial
requests
python-jose[cryptography]
passlib[bcrypt]
python-multipart
